Make an attempt at adding proper error handling.
If something goes wrong, we should be sure to pass an error value as
the first argument to the callback. Errors can be set before the request
fires, or after it returns so we need to handle several cases here. To
be as similar to SuperAgent as possible, we attempt to create a real
Javascript error using the evaluate function.