Add support for reading SSL configuration.
Node.js exposes the tlsSocket where one can get the peer certificate and
protocol information. This should accomplish the same, but in a slightly
different way. We emit the same event 'secureConnect', but instead of
exposing the socket, we add functions on the event to get the cipher,
certificate and protocol.