Extend the support for promises.
It can be useful to want to synchronize several promises and only act
when all of them have been fulfilled. As it turns out there is already a
function for this defined in the various Promise specifications so
implement that function along with the other ones in the spec for
completeness.